Refactoring project struct to packages
