+++ src/fibonacci/model.go
