+++ src/go.mod
@@ -1,0 +1,3 @@
+module github.com/akaKAIN/Go_GB_base/src
+
+go 1.15
