+++ src/mathoperation/mathoperation_test.go
@@ -1,4 +1,4 @@
-package main
+package mathoperation
 
 import "testing"
